+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v0.1.7
+
+...
+
 ## v0.1.6
 
 Added: